INotifyPropertyChanged

+++ DAL/DALImp.cs
@@ -105,8 +105,11 @@ namespace DAL
             using (var ctx = new HistoryDb())
             {
                 var y = (from x in ctx.HistoryFlights where x.Id == IdKey select x).First();
-                ctx.HistoryFlights.Remove(y);
-                ctx.SaveChanges();
+                if (y == null)
+                {
+                    ctx.HistoryFlights.Remove(y);
+                    ctx.SaveChanges();
+                }
 
             }
